fix(striker-ui): hoist api host net list conversion
